% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ants_pipeline.R
\name{ants_pipeline}
\alias{ants_pipeline}
\title{ANTS Analysis Pipeline}
\usage{
ants_pipeline(t1, t2, flair, pd, bias_correction = TRUE, ...)
}
\arguments{
\item{t1}{filename of T1 Image}

\item{t2}{filename of T2 Image}

\item{flair}{filename of FLAIR Image}

\item{pd}{filename of PD Image}

\item{bias_correction}{Should Bias correction be done?}

\item{...}{pass to other stuff}
}
\value{
Stuff
}
\description{
Run my analysis pipeline
}

