% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/copy_data.R
\name{copy_data}
\alias{copy_data}
\title{Copy Data for ENAR SC 2015}
\usage{
copy_data(outdir, get_mipav = TRUE)
}
\arguments{
\item{outdir}{Output directory for data}

\item{get_mipav}{Should it open URL for mipav?}
}
\description{
Copies files from system.file for you
}

