% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dcm2nii.R
\name{dcm2nii}
\alias{dcm2nii}
\title{Dicom to NIfTI wrapper}
\usage{
dcm2nii(path, outfile = NULL, retimg = TRUE, ...)
}
\arguments{
\item{path}{path to DICOM files}

\item{outfile}{output filename}

\item{retimg}{Return Image}

\item{...}{additional options to \code{\link{readDICOM}}}
}
\description{
Converts a directory of DICOMs to NIfTI
}
\examples{
\dontrun{
dicom_path = system.file("FLAIR_DICOM/Brainix", package="ENARSC2015")
img = dcm2nii(dicom_path, retimg = TRUE)
}
}

