% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENARSC2015-package.R
\docType{package}
\name{ENARSC2015-package}
\alias{ENARSC2015-package}
\title{ENAR Short Course in Imaging 2015}
\description{
This package wraps up ENAR Short Course in Imaging 2015 into an
R package for users
}

