% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hox_subcort.R
\docType{data}
\name{hoxsubcort.df}
\alias{hoxsubcort.df}
\title{Labels corresponding to Harvard-Oxford Subcortical Atlas}
\description{
Labels corresponding to Harvard-Oxford Subcortical Atlas
}
\seealso{
\link{hoxsubcort.img}
}
\keyword{data}

